Delete index.html
The index.html file was removed from the project files. This change is part of a larger refactoring process with the goal to migrate the website to a Jekyll-based implementation. As such, the index.html file is no longer necessary as content and layout will now be managed through markdown files and Jekyll templates.