Add jekyll plugins and update site information
Jekyll's SEO Tag and Remote Theme plugins have been added to both the Gemfile and the _config file for website optimization. Updates were also made to site metadata in the _config file, such as adding author information, social links, and a remote theme. The usage of the favicon in custom-head.html was also modified.